+++ routes/user.routes.js
@@ -5,6 +5,6 @@ const user_controller = require('../controllers/user.controller');
 
 router.post('/recover-password', user_controller.recoverPassword);
 router.get('/verify-token', user_controller.verifyToken);
-router.post('/register-password', user_controller.registerPassword);
+router.post('/update-password', user_controller.updatePassword);
 
 module.exports = router;