+++ routes/general.routes.js
@@ -8,6 +8,7 @@ const externalCollabsRoutes = require("./externalCollabs.routes");
 const supplyBatchRouter = require("./supplyBatch.routes")
 
 router.use("/beneficiary", beneficiary_routes);
+// Supplies routes
 router.use("/supplies", suppliesRoutes);
 router.use("/externalCollabs", externalCollabsRoutes);
 router.use("/supplyBatch", supplyBatchRouter);