+++ CHANGELOG.md
@@ -4,6 +4,8 @@
  - Added the business logic in 'WorkshopsController' to generate a new workshop.
  - Added the 'save' and 'add' methods to the 'WorkshopsModel' to interact with the data base.
  - Added the routes in 'WorkshopsRoutes' and 'GeneralRoutes' to access to the model and controller.
+ - Merge with realese 0-1-0
+ - Update supplies model
 
 ## [0.1.0] - 03/11/2025
 
@@ -11,6 +13,12 @@
  - Implemented the endpoint 'DELETE /api/beneficiarios/:id' for the user story 'BEN-004'.
  - Added the business logic in 'BeneficiaryController' to validate the status (can't erase active beneficiaries).
  - Added the 'fetchById' and 'remove' methods to the 'BeneficiaryModel' to interact with the data base.
+
+## [0.0.0] - 02/11/2025 
+
+### Added
+- Updated supplies model to follow database names
+- Implemented routes, controller, and model for the Get Supplies List use case.
 
 ## [0.1.0] - 01/11/2025